move formatting and idempotence tests to `src/test.rs`
I want to allow contributors to configure formatting tests via commetns,
just like you can in rustfmt. In order to do that I want to add internal
helpers to set all that up. In order to do that I need to move these
integration tests into the lib.
